Server: remove SSL_PROTOCOL-test, not supported via Debian.
To get this variable, SSLOptions +CompatEnvVar must be set, however,
this is not longer supported by Apache on Debian, rendering the test
useless (as it will block anyone).

Stronger requirements are therefore placed upon the sysadmin running the
site, to make sure that the protocols used are secure (i.e. only SSLv3
and TLS should be used. With emphasis on TLS).

Signed-off-by: Henrik Austad <henrik@austad.us>